Fix upgrade_lean workflow not triggering build.
Also allow it to be manually triggered.

The problem is described at
https://github.community/t/push-from-action-does-not-trigger-subsequent-action/16854

I explored one other approaches before going with how things are done
in this commit. Approach not taken: Use the `workflow_run` event to have
the `build` workflow run after the `upgrade_lean` workflow. This mostly
works but it gets quite messy conditioning the `build` workflow on
whether the `upgrade_lean` workflow was successful.